feat: add checklist validation for test run and refactor publish validation
