+++ web/app/components/workflow-app/components/workflow-header/app-publisher-trigger.tsx
@@ -3,24 +3,20 @@ import {
   useCallback,
   useMemo,
 } from 'react'
-import { useEdges, useNodes, useStore as useReactflowStore } from 'reactflow'
+import { useStore as useReactflowStore } from 'reactflow'
 import { useTranslation } from 'react-i18next'
 import {
   useStore,
   useWorkflowStore,
 } from '@/app/components/workflow/store'
 import {
-  useChecklist,
   useChecklistBeforePublish,
   useNodesReadOnly,
   useNodesSyncDraft,
+  useWorkflowRunValidation,
 } from '@/app/components/workflow/hooks'
 import AppPublisher from '@/app/components/app/app-publisher'
 import { useFeatures } from '@/app/components/base/features/hooks'
-import type {
-  CommonEdgeType,
-  CommonNodeType,
-} from '@/app/components/workflow/types'
 import {
   BlockEnum,
   InputVarType,
@@ -81,17 +77,13 @@ const AppPublisherTrigger = () => {
   }, [appID, setAppDetail])
 
   const { mutateAsync: publishWorkflow } = usePublishWorkflow(appID!)
-  const nodes = useNodes<CommonNodeType>()
-  const edges = useEdges<CommonEdgeType>()
-  const needWarningNodes = useChecklist(nodes, edges)
+  const { validateBeforeRun } = useWorkflowRunValidation()
 
   const updatePublishedWorkflow = useInvalidateAppWorkflow()
   const onPublish = useCallback(async (params?: PublishWorkflowParams) => {
     // First check if there are any items in the checklist
-    if (needWarningNodes.length > 0) {
-      notify({ type: 'error', message: t('workflow.panel.checklistTip') })
+    if (!validateBeforeRun())
       throw new Error('Checklist has unresolved items')
-    }
 
     // Then perform the detailed validation
     if (await handleCheckBeforePublish()) {
@@ -112,7 +104,7 @@ const AppPublisherTrigger = () => {
     else {
       throw new Error('Checklist failed')
     }
-  }, [needWarningNodes, handleCheckBeforePublish, publishWorkflow, notify, t, updatePublishedWorkflow, appID, updateAppDetail, invalidateAppTriggers, workflowStore, resetWorkflowVersionHistory])
+  }, [validateBeforeRun, handleCheckBeforePublish, publishWorkflow, updatePublishedWorkflow, appID, updateAppDetail, invalidateAppTriggers, workflowStore, resetWorkflowVersionHistory])
 
   const onPublisherToggle = useCallback((state: boolean) => {
     if (state)
